cleaned UI

+++ src/transform_data.py
@@ -6,13 +6,12 @@ to be put into deep learning model
 import numpy as np
 from skimage import transform
 from fastai.vision import Image
-from fastai.torch_core import to_half
-from torch import HalfTensor
+from torch import FloatTensor
 
 
 def fastai_image(img):
     """Turns numpy array into fastai Image object"""
-    img = HalfTensor(img)
+    img = FloatTensor(img)
     img = img.permute(2, 0, 1)
     return Image(img)
